Changed the overview components styles
