--- src/components/AboutUsComponents/Header.jsx
@@ -1,10 +1,0 @@
-
-const Header = () => {
-  return (
-    <div>
-      Header
-    </div>
-  )
-}
-
-export default Header
