+++ src/components/AboutUsComponents/OverviewComponents/Button.jsx
@@ -1,0 +1,12 @@
+
+const Button = ({ title, type, ...props }) => {
+  return (
+    <div>
+      <button type={type} className={props.className}>
+        {title}
+      </button>
+    </div>
+  );
+};
+
+export default Button;
