+++ src/views/AboutUs/AboutUs.jsx
@@ -1,4 +1,4 @@
-import Header from "../../components/AboutUsComponents/Header"
+
 
 const AboutUs = () => {
   return (
